backend/requirements.txt
duckdb
pandas
numpy
scikit-learn
matplotlib
pydantic
langchain
python-dotenv
openai
langchain_community
langchain-openai
groq
langgraph
boto3
redis
fastapi
uvicorn
celery
slowapi